Reorganize fcli files a bit
